Fixed `explore.dart` issue
