--- lib/pages/explore.dart
@@ -1,151 +1,0 @@
-import 'dart:ui';
-
-import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
-import 'package:learn/utils/constants.dart';
-
-// Explore Page
-class ExplorePage extends StatelessWidget {
-  const ExplorePage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return SafeArea(
-        child: CustomScrollView(
-      slivers: [
-        SliverAppBar(
-          title: Padding(
-            padding: const EdgeInsets.fromLTRB(0, 12, 16, 4),
-            child: Text(
-              "Explore", 
-              style: Theme.of(context)
-                  .textTheme
-                  .headlineLarge!
-                  .copyWith(fontWeight: FontWeight.bold, fontSize: 30.0),
-            ),
-          ),
-        ),
-        SliverList(
-          delegate: SliverChildBuilderDelegate(
-            (context, index) {
-              return GestureDetector(
-                onTap: () => Navigator.push(
-                  context,
-                  AppConstants.modules[index].route,
-                ),
-                child: Container(
-                    margin:
-                        const EdgeInsets.symmetric(horizontal: 24, vertical: 12),
-                    height: 200,
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(16),
-                      boxShadow: [
-                        BoxShadow(
-                          color: Colors.black.withOpacity(0.2),
-                          spreadRadius: 2,
-                          blurRadius: 5,
-                          offset: const Offset(0, 3),
-                        ),
-                      ],
-                    ),
-                    child: ClipRRect(
-                      borderRadius: BorderRadius.circular(16),
-                      child: Stack(
-                        fit: StackFit.expand,
-                        alignment: Alignment.center,
-                        children: [
-                          ImageFiltered(
-                            imageFilter: ImageFilter.blur(sigmaX: 5, sigmaY: 5),
-                            child: Image.asset(
-                              AppConstants.modules[index].thumbnailPath,
-                              fit: BoxFit.cover,
-                            ),
-                          ),
-                          Positioned.fill(
-                            child: Align(
-                              child: Column(
-                                mainAxisAlignment: MainAxisAlignment.center,
-                                children: [
-                                  Text(
-                                    AppConstants.modules[index].name,
-                                    style: Theme.of(context)
-                                        .textTheme
-                                        .headlineMedium!
-                                        .copyWith(
-                                      color: Colors.white,
-                                      fontWeight: FontWeight.bold,
-                                      shadows: [
-                                        const Shadow(
-                                          color: Colors.black,
-                                          offset: Offset(2, 1),
-                                          blurRadius: 4,
-                                        ),
-                                      ],
-                                    ),
-                                  ),
-                                  Text(
-                                    AppConstants.modules[index].description,
-                                    style: Theme.of(context)
-                                        .textTheme
-                                        .bodyMedium!
-                                        .copyWith(
-                                      color: Colors.white,
-                                      fontWeight: FontWeight.bold,
-                                      shadows: [
-                                        const Shadow(
-                                          color: Colors.black,
-                                          offset: Offset(2, 1),
-                                          blurRadius: 2,
-                                        ),
-                                      ],
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    )),
-              );
-            },
-            childCount: AppConstants.modules.length,
-          ),
-        ),
-        GestureDetector(
-          onTap: () {
-            Navigator.pushNamed(context, '/quiz');
-          },
-          child: Container(
-            margin: const EdgeInsets.all(5.0),
-            padding: const EdgeInsets.all(8.0),
-            decoration: BoxDecoration(
-              border: Border.all(color: Colors.black, width: 1.0),
-              borderRadius: BorderRadius.circular(8.0),
-              color: Colors.blueAccent,
-            ),
-            child: Row(
-              children: [
-                SizedBox(
-                  width: 50,
-                  height: 50,
-                  child: SvgPicture.asset('assets/explore/notebook.svg'),
-                ),
-                const SizedBox(width: 28.0),
-                const Text(
-                  'Quiz',
-                  style: TextStyle(
-                    fontWeight: FontWeight.bold,
-                    fontSize: 30.0,
-                    fontFamily: 'Comic',
-                    color: Colors.white,
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-      ],
-    ));
-  }
-}
