+++ lib/utils/constants.dart
@@ -1,6 +1,6 @@
 import 'dart:ui';
 
-import '../explore/quiz.dart';
+import '../pages/explore/quiz.dart';
 import 'package:flutter/material.dart';
 import 'package:learn/pages/modules/colours.dart';
 import 'package:learn/pages/modules/parts.dart';
